Update to CEF version 87.1.12+g03f9336+chromium-87.0.4280.88
Add Apple Silicon native build and runtime support (fixes issue #393).

Debug builds on Apple Silicon are currently crashing due to
https://bitbucket.org/chromiumembedded/cef/issues/3070 but Release
builds work fine.

See issue #393 for a quick-start building guide.
