+++ native/CMakeLists.txt
@@ -155,7 +155,7 @@ set(JCEF_SRCS_LINUX
   util_linux.cpp
   util_posix.cpp
   )
-set(JCEF_SRCS_MACOSX
+set(JCEF_SRCS_MAC
   critical_wait_posix.cpp
   signal_restore_posix.cpp
   signal_restore_posix.h
@@ -184,7 +184,7 @@ set(JCEF_HELPER_SRCS
 set(JCEF_HELPER_SRCS_LINUX
   util_posix.cpp
   )
-set(JCEF_HELPER_SRCS_MACOSX
+set(JCEF_HELPER_SRCS_MAC
   util_posix.cpp
   )
 set(JCEF_HELPER_SRCS_WINDOWS
@@ -201,7 +201,7 @@ source_group(jcef FILES ${JCEF_HELPER_SRCS})
 
 # Target binary names.
 set(JCEF_TARGET "jcef")
-if(OS_MACOSX)
+if(OS_MAC)
   set(JCEF_APP_NAME "jcef_app")
   set(JCEF_HELPER_TARGET "jcef_Helper")
   set(JCEF_HELPER_OUTPUT_NAME "jcef Helper")
@@ -285,7 +285,7 @@ endif()
 # Mac OS X configuration.
 #
 
-if(OS_MACOSX)
+if(OS_MAC)
   # Avoid CMP0042 policy errors.
   set(CMAKE_MACOSX_RPATH 1)
