+++ native/browser_process_handler.cpp
@@ -38,22 +38,6 @@ void BrowserProcessHandler::OnContextInitialized() {
   JNI_CALL_VOID_METHOD(env, handle_, "onContextInitialized", "()V");
 }
 
-void BrowserProcessHandler::OnRenderProcessThreadCreated(
-    CefRefPtr<CefListValue> extra_info) {
-  int idx = 0;
-  static std::set<CefMessageRouterConfig, cmpCfg>::iterator iter;
-
-  // Delegate creation of the renderer-side router for query handling.
-  base::AutoLock lock_scope(router_cfg_lock_);
-  for (iter = router_cfg_.begin(); iter != router_cfg_.end(); ++iter) {
-    CefRefPtr<CefDictionaryValue> dict = CefDictionaryValue::Create();
-    dict->SetString("js_query_function", iter->js_query_function);
-    dict->SetString("js_cancel_function", iter->js_cancel_function);
-    extra_info->SetDictionary(idx, dict);
-    idx++;
-  }
-}
-
 CefRefPtr<CefPrintHandler> BrowserProcessHandler::GetPrintHandler() {
   CefRefPtr<CefPrintHandler> result;
   ScopedJNIEnv env;
@@ -85,12 +69,36 @@ void BrowserProcessHandler::OnScheduleMessagePumpWork(int64 delay_ms) {
                        delay_ms);
 }
 
+// static
+CefRefPtr<CefListValue> BrowserProcessHandler::GetMessageRouterConfigs() {
+  int idx = 0;
+  static std::set<CefMessageRouterConfig, cmpCfg>::iterator iter;
+
+  base::AutoLock lock_scope(router_cfg_lock_);
+  if (router_cfg_.empty())
+    return nullptr;
+
+  // Configuration pased to CefHelperApp::OnBrowserCreated.
+  auto router_configs = CefListValue::Create();
+  for (iter = router_cfg_.begin(); iter != router_cfg_.end(); ++iter) {
+    CefRefPtr<CefDictionaryValue> dict = CefDictionaryValue::Create();
+    dict->SetString("js_query_function", iter->js_query_function);
+    dict->SetString("js_cancel_function", iter->js_cancel_function);
+    router_configs->SetDictionary(idx, dict);
+    idx++;
+  }
+
+  return router_configs;
+}
+
+// static
 void BrowserProcessHandler::AddMessageRouterConfig(
     const CefMessageRouterConfig& cfg) {
   base::AutoLock lock_scope(router_cfg_lock_);
   router_cfg_.insert(cfg);
 }
 
+// static
 void BrowserProcessHandler::RemoveMessageRouterConfig(
     const CefMessageRouterConfig& cfg) {
   base::AutoLock lock_scope(router_cfg_lock_);
