+++ native/browser_process_handler.h
@@ -33,11 +33,10 @@ class BrowserProcessHandler : public CefBrowserProcessHandler {
   virtual ~BrowserProcessHandler();
 
   void OnContextInitialized() OVERRIDE;
-  void OnRenderProcessThreadCreated(
-      CefRefPtr<CefListValue> extra_info) OVERRIDE;
   CefRefPtr<CefPrintHandler> GetPrintHandler() OVERRIDE;
   void OnScheduleMessagePumpWork(int64 delay_ms) OVERRIDE;
 
+  static CefRefPtr<CefListValue> GetMessageRouterConfigs();
   static void AddMessageRouterConfig(const CefMessageRouterConfig& cfg);
   static void RemoveMessageRouterConfig(const CefMessageRouterConfig& cfg);
