+++ native/display_handler.cpp
@@ -6,6 +6,66 @@
 
 #include "jni_util.h"
 
+namespace {
+
+int GetCursorId(cef_cursor_type_t type) {
+  ScopedJNIEnv env;
+  if (!env)
+    return 0;
+
+  ScopedJNIClass cls(env, "java/awt/Cursor");
+  if (!cls)
+    return 0;
+
+  JNI_STATIC_DEFINE_INT_RV(env, cls, CROSSHAIR_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, DEFAULT_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, E_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, HAND_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, MOVE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, N_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, NE_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, NW_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, S_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, SE_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, SW_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, TEXT_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, W_RESIZE_CURSOR, 0);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, WAIT_CURSOR, 0);
+
+  switch (type) {
+    case CT_CROSS:
+      return JNI_STATIC(CROSSHAIR_CURSOR);
+    case CT_HAND:
+      return JNI_STATIC(HAND_CURSOR);
+    case CT_IBEAM:
+      return JNI_STATIC(TEXT_CURSOR);
+    case CT_WAIT:
+      return JNI_STATIC(WAIT_CURSOR);
+    case CT_EASTRESIZE:
+      return JNI_STATIC(E_RESIZE_CURSOR);
+    case CT_NORTHRESIZE:
+      return JNI_STATIC(N_RESIZE_CURSOR);
+    case CT_NORTHEASTRESIZE:
+      return JNI_STATIC(NE_RESIZE_CURSOR);
+    case CT_NORTHWESTRESIZE:
+      return JNI_STATIC(NW_RESIZE_CURSOR);
+    case CT_SOUTHRESIZE:
+      return JNI_STATIC(S_RESIZE_CURSOR);
+    case CT_SOUTHEASTRESIZE:
+      return JNI_STATIC(SE_RESIZE_CURSOR);
+    case CT_SOUTHWESTRESIZE:
+      return JNI_STATIC(SW_RESIZE_CURSOR);
+    case CT_WESTRESIZE:
+      return JNI_STATIC(W_RESIZE_CURSOR);
+    case CT_MOVE:
+      return JNI_STATIC(MOVE_CURSOR);
+    default:
+      return JNI_STATIC(DEFAULT_CURSOR);
+  }
+}
+
+}  // namespace
+
 DisplayHandler::DisplayHandler(JNIEnv* env, jobject handler)
     : handle_(env, handler) {}
 
@@ -106,6 +166,26 @@ bool DisplayHandler::OnConsoleMessage(CefRefPtr<CefBrowser> browser,
       "Ljava/lang/String;Ljava/lang/String;I)Z",
       Boolean, jreturn, jbrowser.get(), jlevel, jmessage.get(), jsource.get(),
       line);
+
+  return (jreturn != JNI_FALSE);
+}
+
+// TODO(JCEF): Expose all parameters.
+bool DisplayHandler::OnCursorChange(CefRefPtr<CefBrowser> browser,
+                                    CefCursorHandle cursor,
+                                    cef_cursor_type_t type,
+                                    const CefCursorInfo& custom_cursor_info) {
+  ScopedJNIEnv env;
+  if (!env)
+    return false;
+
+  ScopedJNIBrowser jbrowser(env, browser);
+  const int cursorId = GetCursorId(type);
+  jboolean jreturn = JNI_FALSE;
+
+  JNI_CALL_METHOD(env, handle_, "onCursorChange",
+                  "(Lorg/cef/browser/CefBrowser;I)Z", Boolean, jreturn,
+                  jbrowser.get(), cursorId);
 
   return (jreturn != JNI_FALSE);
 }
